fix(resource dependencies): update terraform and ARM dependencies
idam-api, idam-web-{public/admin} all fail on the first and second terraform execution when using
this module due to incorrect dependencies. By using the resource reference instead of the local,
this enforces implicit dependencies for the rg resolving ResourceGroupNotFound errors and by adding
the dependsOn statement for the serverfarm deployment in the ARM template, this removes the 'Cannot
find ServerFarm with name ...' error.